CS-E4660/research/requirements.txt
# Core MCP and AI libraries
mcp>=0.9.0
anthropic>=0.39.0
python-dotenv>=1.0.0

# LangChain for Arcee Agent approach
langchain>=0.1.0
langchain-community>=0.0.20

# Transformers and ML libraries
transformers>=4.36.0
torch>=2.1.0
accelerate>=0.25.0

# Optional: vLLM for production (requires GPU)
# vllm>=0.2.0

# Optional: Quantization support
# bitsandbytes>=0.41.0

# API and web frameworks
fastapi>=0.108.0
uvicorn>=0.25.0
httpx>=0.25.0
pydantic>=2.0.0

# AWS and SageMaker (for Approach C)
boto3>=1.34.0
sagemaker>=2.200.0

# Utilities
aiofiles>=23.2.0
asyncio>=3.4.3
